# api/assets/openapi.json
{
	"openapi": "3.0.0",
	"servers": [
		{
			"url": "https://api.copycord.com/api/v{version}",
			"description": "Official copycord instance",
			"variables": {
				"version": {
					"default": "9",
					"enum": ["8", "9"]
				}
			}
		}
	],
	"info": {
		"description": "copycord is a free open source selfhostable discord compatible chat, voice and video platform",
		"version": "1.0.0",
		"title": "copycord HTTP API Routes",
		"termsOfService": "",
		"contact": {
			"name": "copycord"
		},
		"license": {
			"name": "AGPLV3",
			"url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
		}
	},
...
	"paths": {
...
		"/users/@me/channels/": {
			"get": {
				"description": "",
				"responses": {
					"default": {
						"description": "not documented"
					}
				},
				"tags": ["users"]
			},
			"post": {
				"description": "",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DmChannelCreateSchema"
							}
						}
					}
				},
				"responses": {
					"default": {
						"description": "not documented"
					}
				},
				"tags": ["users"]
			}
		},
...
		"/invites/{code}": {
			"get": {
				"description": "",
				"responses": {
					"default": {
						"description": "not documented"
					}
				},
				"parameters": [
					{
						"name": "code",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "code"
					}
				],
				"tags": ["invites"]
			},
			"post": {
				"description": "",
				"responses": {
					"default": {
						"description": "not documented"
					}
				},
				"parameters": [
					{
						"name": "code",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "code"
					}
				],
				"tags": ["invites"]
			},
			"delete": {
				"description": "",
				"responses": {
					"default": {
						"description": "not documented"
					}
				},
				"parameters": [
					{
						"name": "code",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "code"
					}
				],
				"tags": ["invites"]
			}
		},
...
		"/guilds/{guild_id}/members/": {
			"get": {
				"description": "",
				"responses": {
					"default": {
						"description": "not documented"
					}
				},
				"parameters": [
					{
						"name": "guild_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "guild_id"
					}
				],
				"tags": ["guilds"]
			}
		},
		"/guilds/{guild_id}/members/{member_id}/nick/": {
			"patch": {
				"description": "",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MemberNickChangeSchema"
							}
						}
					}
				},
				"responses": {
					"default": {
						"description": "not documented"
					}
				},
				"parameters": [
					{
						"name": "guild_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "guild_id"
					},
					{
						"name": "member_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "member_id"
					}
				],
				"tags": ["guilds"]
			}
		},
		"/guilds/{guild_id}/members/{member_id}/": {
			"get": {
				"description": "",
				"responses": {
					"default": {
						"description": "not documented"
					}
				},
				"parameters": [
					{
						"name": "guild_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "guild_id"
					},
					{
						"name": "member_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "member_id"
					}
				],
				"tags": ["guilds"]
			},
			"patch": {
				"description": "",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MemberChangeSchema"
							}
						}
					}
				},
				"responses": {
					"default": {
						"description": "not documented"
					}
				},
				"parameters": [
					{
						"name": "guild_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "guild_id"
					},
					{
						"name": "member_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "member_id"
					}
				],
				"tags": ["guilds"]
			},
			"put": {
				"description": "",
				"responses": {
					"default": {
						"description": "not documented"
					}
				},
				"parameters": [
					{
						"name": "guild_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "guild_id"
					},
					{
						"name": "member_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "member_id"
					}
				],
				"tags": ["guilds"]
			},
			"delete": {
				"description": "##### Requires the ``KICK_MEMBERS`` permission\n",
				"responses": {
					"default": {
						"description": "not documented"
					}
				},
				"parameters": [
					{
						"name": "guild_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "guild_id"
					},
					{
						"name": "member_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "member_id"
					}
				],
				"tags": ["guilds"]
			}
		},
		"/guilds/{guild_id}/invites/": {
			"get": {
				"description": "##### Requires the ``MANAGE_GUILD`` permission\n",
				"responses": {
					"default": {
						"description": "not documented"
					}
				},
				"parameters": [
					{
						"name": "guild_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "guild_id"
					}
				],
				"tags": ["guilds"]
			}
		},
...
		"/guilds/{guild_id}/channels/": {
			"get": {
				"description": "",
				"responses": {
					"default": {
						"description": "not documented"
					}
				},
				"parameters": [
					{
						"name": "guild_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "guild_id"
					}
				],
				"tags": ["guilds"]
			},
			"post": {
				"description": "##### Requires the ``MANAGE_CHANNELS`` permission\n",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChannelModifySchema"
							}
						}
					}
				},
				"responses": {
					"default": {
						"description": "not documented"
					}
				},
				"parameters": [
					{
						"name": "guild_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "guild_id"
					}
				],
				"tags": ["guilds"]
			},
			"patch": {
				"description": "##### Requires the ``MANAGE_CHANNELS`` permission\n",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChannelReorderSchema"
							}
						}
					}
				},
				"responses": {
					"default": {
						"description": "not documented"
					}
				},
				"parameters": [
					{
						"name": "guild_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "guild_id"
					}
				],
				"tags": ["guilds"]
			}
		},
...
		"/channels/{channel_id}/invites/": {
			"post": {
				"description": "##### Requires the ``CREATE_INSTANT_INVITE`` permission\n",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/InviteCreateSchema"
							}
						}
					}
				},
				"responses": {
					"default": {
						"description": "not documented"
					}
				},
				"parameters": [
					{
						"name": "channel_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "channel_id"
					}
				],
				"tags": ["channels"]
			},
			"get": {
				"description": "##### Requires the ``MANAGE_CHANNELS`` permission\n",
				"responses": {
					"default": {
						"description": "not documented"
					}
				},
				"parameters": [
					{
						"name": "channel_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "channel_id"
					}
				],
				"tags": ["channels"]
			}
		},
		"/channels/{channel_id}/": {
			"get": {
				"description": "##### Requires the ``VIEW_CHANNEL`` permission\n",
				"responses": {
					"default": {
						"description": "not documented"
					}
				},
				"parameters": [
					{
						"name": "channel_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "channel_id"
					}
				],
				"tags": ["channels"]
			},
			"delete": {
				"description": "##### Requires the ``MANAGE_CHANNELS`` permission\n",
				"responses": {
					"default": {
						"description": "not documented"
					}
				},
				"parameters": [
					{
						"name": "channel_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "channel_id"
					}
				],
				"tags": ["channels"]
			},
			"patch": {
				"description": "##### Requires the ``MANAGE_CHANNELS`` permission\n",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChannelModifySchema"
							}
						}
					}
				},
				"responses": {
					"default": {
						"description": "not documented"
					}
				},
				"parameters": [
					{
						"name": "channel_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "channel_id"
					}
				],
				"tags": ["channels"]
			}
		},
...
	},
...
	"components": {
		"schemas": {
...
			"ChannelPermissionOverwriteType": {
				"enum": [0, 1],
				"type": "number"
			},
...
			"DmChannelCreateSchema": {
				"type": "object",
				"properties": {
					"name": {
						"type": "string"
					},
					"recipients": {
						"type": "array",
						"items": {
							"type": "string"
						}
					}
				},
				"required": ["recipients"]
			},
			"ChannelModifySchema": {
				"type": "object",
				"properties": {
					"name": {
						"maxLength": 100,
						"type": "string"
					},
					"type": {
						"enum": [0, 1, 10, 11, 12, 13, 2, 3, 4, 5, 6],
						"type": "number"
					},
					"topic": {
						"type": "string"
					},
					"icon": {
						"type": "string",
						"nullable": true
					},
					"bitrate": {
						"type": "integer"
					},
					"user_limit": {
						"type": "integer"
					},
					"rate_limit_per_user": {
						"type": "integer"
					},
					"position": {
						"type": "integer"
					},
					"permission_overwrites": {
						"type": "array",
						"items": {
							"type": "object",
							"properties": {
								"id": {
									"type": "string"
								},
								"type": {
									"$ref": "#/components/schemas/ChannelPermissionOverwriteType"
								},
								"allow": {
									"type": "number"
								},
								"deny": {
									"type": "number"
								}
							},
							"additionalProperties": false,
							"required": ["allow", "deny", "id", "type"]
						}
					},
					"parent_id": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"nsfw": {
						"type": "boolean"
					},
					"rtc_region": {
						"type": "string"
					},
					"default_auto_archive_duration": {
						"type": "integer"
					}
				}
			},
...
			"InviteCreateSchema": {
				"type": "object",
				"properties": {
					"target_user_id": {
						"type": "string"
					},
					"target_type": {
						"type": "string"
					},
					"validate": {
						"type": "string"
					},
					"max_age": {
						"type": "integer"
					},
					"max_uses": {
						"type": "integer"
					},
					"temporary": {
						"type": "boolean"
					},
					"unique": {
						"type": "boolean"
					},
					"target_user": {
						"type": "string"
					},
					"target_user_type": {
						"type": "integer"
					}
				}
			},
...
			"MemberNickChangeSchema": {
				"type": "object",
				"properties": {
					"nick": {
						"type": "string"
					}
				},
				"required": ["nick"]
			},
			"MemberChangeSchema": {
				"type": "object",
				"properties": {
					"roles": {
						"type": "array",
						"items": {
							"type": "string"
						}
					}
				}
			},
...
			"ChannelReorderSchema": {
				"type": "array",
				"items": {
					"type": "object",
					"properties": {
						"id": {
							"type": "string"
						},
						"position": {
							"type": "integer"
						},
						"lock_permissions": {
							"type": "boolean"
						},
						"parent_id": {
							"type": "string"
						}
					},
					"additionalProperties": false,
					"required": ["id"]
				}
			},
...
		},
...
		"securitySchemes": {
			"Token": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		},
...
	}
}